refactor: share chunkMessageTo64Bytes; drop NFTTab char-based slicer
- Move chunkMessageTo64Bytes cslTools.js -> utils.js (pure fn; Buffer.byteLength
  instead of TextEncoder so it also runs under the jsdom test env)
- cip20Tab imports it from utils
- NFTTab: replace char-based sliceBy64Char (could exceed 64 bytes on multibyte
  text) with a thin adapter over the shared byte-based util, keeping CIP-25's
  string-or-array output
- Add chunkMessageTo64Bytes tests
