+++ src/utils/utils.js
@@ -8,6 +8,29 @@ export function hexToBytes(hex) {
   return Buffer.from(hex, 'hex')
 }
 
+// Splits a string into an array of chunks each <= 64 bytes when UTF-8 encoded.
+// Cardano metadatum text strings (CIP-20 messages, CIP-25 NFT fields) are capped
+// at 64 BYTES (not chars). We accumulate whole code points (iterating a string
+// yields code points, not UTF-16 units) so we never split a multibyte character.
+export function chunkMessageTo64Bytes(message) {
+  const chunks = []
+  let current = ''
+  let currentBytes = 0
+  for (const ch of message) {
+    const chBytes = Buffer.byteLength(ch, 'utf8')
+    if (currentBytes + chBytes > 64) {
+      if (current) chunks.push(current)
+      current = ch
+      currentBytes = chBytes
+    } else {
+      current += ch
+      currentBytes += chBytes
+    }
+  }
+  if (current) chunks.push(current)
+  return chunks
+}
+
 export function wasmMultiassetToJSONs(wasmMultiasset) {
   let assetValue = []
   const wasmScriptHashes = wasmMultiasset?.keys()
